+++ test/fixtures/typescript/decorators.ts
@@ -1,19 +1,33 @@
-function decorator (...args: any) {
-  console.log('Decorator called with arguments:', args)
+import "reflect-metadata";
+function decorator(...args: any) {
+  console.log("Decorator called with " + args.length + " arguments.");
+}
+
+function anotherDecorator() {
+  return function (object: any, propertyName: any) {
+    console.log(
+      "Decorator metadata keys: " +
+        Reflect.getMetadataKeys(object, propertyName)
+    );
+  };
 }
 
 @decorator
 export default class DecoratedClass {
-  @decorator
-    decoratedProperty = null
+  @anotherDecorator()
+  decoratedProperty: string;
 
   @decorator
-  get decoratedAccessor () {
-    return null
+  get decoratedAccessor() {
+    return null;
   }
 
   @decorator
-  decoratedFunction (@decorator decoratedParameter: any) {
-    return decoratedParameter
+  decoratedFunction(@decorator decoratedParameter: any) {
+    return decoratedParameter;
+  }
+
+  constructor() {
+    this.decoratedProperty = "foo";
   }
 }